open the workspace apps launcher as a horizontal titlebar dropdown
