+++ packages/shared/types.ts
@@ -184,7 +184,6 @@ export type IpcMainEvents =
         app: SupportedWorkspaceApp,
         openBehavior?: WorkspaceAppOpenBehavior,
       ];
-      "workspaceApps.showLauncherMenu": [];
       "doNotDisturb.toggle": [];
       "doNotDisturb.showOptions": [];
       "downloads.toggleRecentDownloadHistoryPopup": [];
